fix(rust/local_core): times_ms 的 Windows 分支编译不过
std::os::windows::fs::MetadataExt 的 last_write_time/creation_time 返回
FILETIME（u64，100ns 计数），而 times_ms 的签名是 (i64, i64)，因此
`cargo build --target x86_64-pc-windows-msvc` 报 E0308。macOS 走
`#[cfg(unix)]` 那条（mtime() 本身就是 i64），所以这个分支从未被编译过。

减法口径不变（FILETIME/10000 再减去 1601→1970 的毫秒数，与 std 内部一致），
只把转换放在减法之前：文件系统不支持时间时 FILETIME 为 0，按 u64 减常量会
下溢，转成 i64 后得到一个负数哨兵而不是 panic。

+++ rust/local_core/src/file_ops/execute.rs
@@ -502,8 +502,9 @@ fn times_ms(metadata: &fs::Metadata) -> (i64, i64) {
     #[cfg(windows)]
     {
         use std::os::windows::fs::MetadataExt;
-        let mtime = metadata.last_write_time() / 10_000 - 11_644_473_600_000;
-        let ctime = metadata.creation_time() / 10_000 - 11_644_473_600_000;
+        // 先转 i64 再做减法：文件系统不支持时间时 FILETIME 为 0，按 u64 减常量会下溢。
+        let mtime = metadata.last_write_time() as i64 / 10_000 - 11_644_473_600_000;
+        let ctime = metadata.creation_time() as i64 / 10_000 - 11_644_473_600_000;
         (mtime, ctime)
     }
     #[cfg(not(any(unix, windows)))]
